# docs/openapi/platform-openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Observability Platform API",
    "description": "File n�y du?c g?p th? c�ng t? /v3/api-docs c?a ingestion-service v� core-app t?i th?i di?m t?o � c� th? l?ch n?u code sau n�y d?i m� kh�ng regenerate l?i; ngu?n s? th?t lu�n l� 2 endpoint /v3/api-docs s?ng.",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/v1/dlq/process": {
      "post": {
        "tags": [
          "dlq-controller"
        ],
        "operationId": "processDlqMessages",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DlqProcessRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UnifiedResponseString"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/dlq": {
      "get": {
        "tags": [
          "dlq-controller"
        ],
        "operationId": "getDlqMessages",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UnifiedResponseListDlqMessageDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/telemetry/logs": {
      "post": {
        "tags": [
          "Log Ingestion"
        ],
        "summary": "Ingest a single log event",
        "description": "Accepts a single log event. Duplicate `eventId`s are accepted (202) but not forwarded to Kafka. Sensitive data in specific fields may be masked.",
        "operationId": "ingestSingleLog",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogIngestionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Log event accepted for processing",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseString"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or revoked API Key",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded for this service",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Message broker unavailable",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "X-API-Key": []
          }
        ]
      }
    },
    "/api/v1/telemetry/logs/batch": {
      "post": {
        "tags": [
          "Log Ingestion"
        ],
        "summary": "Ingest a batch of log events",
        "description": "Accepts multiple log events in a single request. Rate limit applies per request, not per event inside the batch.",
        "operationId": "ingestBatchLogs",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogBatchIngestionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Batch log events accepted for processing",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListString"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or revoked API Key",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded for this service",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Message broker unavailable",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "X-API-Key": []
          }
        ]
      }
    },
    "/api/v1/telemetry/logs/batch/raw": {
      "post": {
        "tags": [
          "Log Ingestion"
        ],
        "summary": "Ingest a raw JSON array of log events (Agent-friendly)",
        "description": "Accepts multiple log events as a raw JSON array. Used by log shippers like Fluent Bit. Rate limit applies per request, not per event inside the batch.",
        "operationId": "ingestRawBatchLogs",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/LogIngestionRequest"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Batch log events accepted for processing",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListString"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or revoked API Key",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded for this service",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Message broker unavailable",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "X-API-Key": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "DlqProcessRequest": {
        "required": [
          "action"
        ],
        "type": "object",
        "properties": {
          "action": {
            "type": "string"
          },
          "limit": {
            "maximum": 100,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "UnifiedResponseString": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "Status code of the response (e.g., SUCCESS, VALIDATION_ERROR, NOT_FOUND)",
            "example": "SUCCESS"
          },
          "data": {
            "type": "string",
            "description": "The main response data payload"
          },
          "message": {
            "type": "string",
            "description": "Message providing more context, usually on errors"
          },
          "timestamp": {
            "type": "string",
            "description": "Timestamp of the response (ISO-8601)",
            "example": "2026-09-08T10:15:30Z"
          },
          "path": {
            "type": "string",
            "description": "API path that was called",
            "example": "/api/v1/services"
          },
          "fieldErrors": {
            "type": "array",
            "description": "List of validation errors if status is VALIDATION_ERROR",
            "items": {
              "type": "object",
              "additionalProperties": {
                "type": "string",
                "description": "List of validation errors if status is VALIDATION_ERROR"
              },
              "description": "List of validation errors if status is VALIDATION_ERROR"
            }
          }
        },
        "description": "Unified envelope response format for all API calls"
      },
...
      "DlqMessageDto": {
        "type": "object",
        "properties": {
          "partition": {
            "type": "integer",
            "format": "int32"
          },
          "offset": {
            "type": "integer",
            "format": "int64"
          },
          "key": {
            "type": "string"
          },
          "payload": {
            "type": "string"
          },
          "headers": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "errorReason": {
            "type": "string"
          },
          "firstFailedAt": {
            "type": "string"
          }
        },
        "description": "The main response data payload"
      },
      "UnifiedResponseListDlqMessageDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "Status code of the response (e.g., SUCCESS, VALIDATION_ERROR, NOT_FOUND)",
            "example": "SUCCESS"
          },
          "data": {
            "type": "array",
            "description": "The main response data payload",
            "items": {
              "$ref": "#/components/schemas/DlqMessageDto"
            }
          },
          "message": {
            "type": "string",
            "description": "Message providing more context, usually on errors"
          },
          "timestamp": {
            "type": "string",
            "description": "Timestamp of the response (ISO-8601)",
            "example": "2026-09-08T10:15:30Z"
          },
          "path": {
            "type": "string",
            "description": "API path that was called",
            "example": "/api/v1/services"
          },
          "fieldErrors": {
            "type": "array",
            "description": "List of validation errors if status is VALIDATION_ERROR",
            "items": {
              "type": "object",
              "additionalProperties": {
                "type": "string",
                "description": "List of validation errors if status is VALIDATION_ERROR"
              },
              "description": "List of validation errors if status is VALIDATION_ERROR"
            }
          }
        },
        "description": "Unified envelope response format for all API calls"
      },
...
      "LogIngestionRequest": {
        "required": [
          "level",
          "message",
          "timestamp"
        ],
        "type": "object",
        "properties": {
          "eventId": {
            "type": "string",
            "description": "Unique event identifier for idempotency",
            "example": "evt-12345-abcde"
          },
          "timestamp": {
            "type": "string",
            "description": "Timestamp of the log event (ISO-8601)",
            "format": "date-time",
            "example": "2026-09-08T10:15:30Z"
          },
          "level": {
            "type": "string",
            "description": "Log level (e.g., INFO, ERROR, DEBUG)",
            "example": "ERROR"
          },
          "message": {
            "type": "string",
            "description": "The main log message",
            "example": "Database connection timed out after 30s"
          },
          "traceId": {
            "type": "string",
            "description": "Trace ID for distributed tracing correlation",
            "example": "4bf92f3577b34da6a3ce929d0e0e4736"
          },
          "spanId": {
            "type": "string",
            "description": "Span ID for distributed tracing correlation",
            "example": "00f067aa0ba902b7"
          },
          "host": {
            "type": "string",
            "description": "Host name or identifier where the log originated"
          },
          "instanceId": {
            "type": "string",
            "description": "Instance ID for clustered environments"
          },
          "logger": {
            "type": "string",
            "description": "Logger name, usually the class path"
          },
          "httpMethod": {
            "type": "string",
            "description": "HTTP method if applicable",
            "example": "GET"
          },
          "endpoint": {
            "type": "string",
            "description": "API endpoint path",
            "example": "/api/v1/users"
          },
          "statusCode": {
            "type": "integer",
            "description": "HTTP status code",
            "format": "int32",
            "example": 500
          },
          "durationMs": {
            "type": "integer",
            "description": "Execution duration in milliseconds",
            "format": "int64",
            "example": 300
          },
          "metadata": {
            "type": "object",
            "additionalProperties": {
              "type": "object",
              "description": "Additional structured metadata attributes"
            },
            "description": "Additional structured metadata attributes"
          }
        },
        "description": "Request payload for ingesting a single log event"
      },
      "FieldError": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ValidationErrorResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "path": {
            "type": "string"
          },
          "traceId": {
            "type": "string"
          },
          "fieldErrors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "path": {
            "type": "string"
          },
          "traceId": {
            "type": "string"
          }
        }
      },
      "ApiResponseString": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "LogBatchIngestionRequest": {
        "required": [
          "logs"
        ],
        "type": "object",
        "properties": {
          "logs": {
            "maxItems": 100,
            "minItems": 0,
            "type": "array",
            "description": "List of log events (max 100 per batch)",
            "items": {
              "$ref": "#/components/schemas/LogIngestionRequest"
            }
          }
        },
        "description": "Request payload for ingesting a batch of log events"
      },
      "ApiResponseListString": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT Authorization header using the Bearer scheme.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "X-API-Key": {
        "type": "apiKey",
        "description": "API Key authentication. Required for all endpoints.",
        "name": "X-API-Key",
        "in": "header"
      }
    }
  }
}